backend/pyproject.toml
[tool.poetry]
name = "school-management-backend"
version = "0.1.0"
description = "Backend API for School Management System"
authors = ["OpenHands <openhands@all-hands.dev>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.5.2"}
alembic = "^1.12.1"
python-multipart = "^0.0.6"
redis = "^5.0.1"
celery = "^5.3.6"
flower = "^2.0.1"
pandas = "^2.1.4"
openpyxl = "^3.1.2"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"
aiofiles = "^23.2.1"
boto3 = "^1.34.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
httpx = "^0.25.2"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3